Add files via upload
Removed most projects for brevity. The final project I did in under a week and I am well aware it is not perfect, you can read the comments. 

After skimming, project 7 I felt had something interesting to show with a 2D array in JS, but most of the projects were nothing too interesting. A lot of basic pages and JQuery.