Add files via upload
The final project covered a lot of the topics throughout the course. I removed everything else as they were redundant.

The pictures, Visio, PDFs, and word docs are part of the project. The goal was very general; create a database with relations and cover course topics. Looking at the relationship charts, you can see it started a little over-complicated, but after talking to the teacher about it he helped make the scope more understandable and we simplified it.